+++ K8sUtils/public/Get-PodByJobName.ps1
@@ -15,13 +15,16 @@ An example
 One or more pods for the job
 #>
 function Get-PodByJobName {
+    [CmdletBinding()]
     param (
         [string] $JobName,
         [string] $Namespace = "default"
     )
 
+    Write-Verbose "kubectl get job $JobName --namespace $Namespace -o json"
     $job = kubectl get job $JobName --namespace $Namespace -o json | ConvertFrom-Json
     if ($job) {
+        Write-Verbose "kubectl get pod --namespace $Namespace --selector 'batch.kubernetes.io/controller-uid=$($job.metadata.labels.'batch.kubernetes.io/controller-uid')' -o json"
         (kubectl get pod --namespace $Namespace --selector "batch.kubernetes.io/controller-uid=$($job.metadata.labels.'batch.kubernetes.io/controller-uid')" -o json | ConvertFrom-Json).items
     } else {
         Write-Warning "Job $JobName not found in namespace $Namespace"