+++ Tools/MinimalDeploy.tests.ps1
@@ -202,12 +202,28 @@ Describe "Deploys Minimal API" {
         }
     } -Tag 'Sad','t24'
 
+    It "tests no changes" {
+        $deploy1 = Deploy-Minimal -PassThru -SkipPreHook -SkipInit -TimeoutSecs 10 -SkipSetStartTime
+
+        Test-Deploy $deploy1
+
+        Test-MainPod $deploy1.PodStatuses[0]
+
+        $deploy2 = Deploy-Minimal -PassThru -SkipPreHook -SkipInit -TimeoutSecs 10 -SkipSetStartTime
+
+        Test-Deploy $deploy2
+
+        Test-MainPod $deploy2.PodStatuses[0]
+
+        $deploy1.PodStatuses[0].PodName | Should -Be $deploy2.PodStatuses[0].PodName
+    } -Tag 'Happy','t25'
+    
     It "tests rollback if uninstalled" {
         helm uninstall test
         $deploy = Deploy-Minimal -PassThru -SkipInit -SkipPreHook -Fail
         Test-Deploy $deploy -Running $false -RollbackStatus 'RolledBack'
 
         Test-MainPod $deploy.PodStatuses[0] -status 'Crash'
-   } -Tag 'Sad','t25'
+   } -Tag 'Sad','t26'
 }